unify all tests idiomatically
